[tflitefile_tools] Move old ir files to parser/tflite/
Let's move old ir files to 'parser/tflite/'. New IRs will be uploaded
soon.
