+++ tools/tflitefile_tool/saver/config_saver.py
@@ -14,7 +14,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from ir.operator_wrapping import Operator
 from printer.string_builder import GetStringShape
